Bug Fixes and RGB Compare
Fixed a few bugs in some files.
Updated to compare for now all files with the RGB, improved duplicate detection a good amount. First by determining if it's a possible match, 3 spots are stored for each picture and compared. If they all match, then a full RGB compare between the two happens to determine duplicate. May in another change make an option to compare this way versus metaData only...